+++ pdata/pmetric/pmetricotlp/metrics.go
@@ -122,9 +122,6 @@ type GRPCClient interface {
 	Export(ctx context.Context, request Request, opts ...grpc.CallOption) (Response, error)
 }
 
-// Deprecated: [0.61.0] Use GRPCClient instead
-type Client = GRPCClient
-
 type metricsClient struct {
 	rawClient otlpcollectormetrics.MetricsServiceClient
 }
@@ -147,9 +144,6 @@ type GRPCServer interface {
 	// alive for the entire life of the application.
 	Export(context.Context, Request) (Response, error)
 }
-
-// Deprecated: [0.61.0] Use GRPCServer instead
-type Server = GRPCServer
 
 // RegisterGRPCServer registers the GRPCServer to the grpc.Server.
 func RegisterGRPCServer(s *grpc.Server, srv GRPCServer) {